pdaoa-prevDate-health-check/requirements.txt
google-api-python-client
google-cloud-monitoring
google-cloud-bigquery[pandas,pyarrow]==1.7.0
pandas-gbq==0.7.0